refactor: changes state management (redux to react-query)

+++ src/data/usecases/_game/_delete-game/index.ts
@@ -1,25 +1,40 @@
-import { IDeleteGameUsecase } from '@domain/usecases';
+import { ForbiddenError, NotFoundError } from '@domain/errors';
+import { IDeleteGameUsecase, IGetGameUsecase } from '@domain/usecases';
 
 import { IGameDAO } from '@data/dao';
-import { DeleteGameObserver } from '@data/observers';
+import { SessionGetterProtocol } from '@data/protocols';
 
 export class DeleteGameUsecase implements IDeleteGameUsecase {
-  private readonly deleteGamePublisher: DeleteGameObserver.Publisher;
+  private readonly sessionGetter: SessionGetterProtocol;
+  private readonly getGame: IGetGameUsecase;
   private readonly gameDAO: IGameDAO;
-
-  public constructor({ deleteGamePublisher, gameDAO }: Deps) {
-    this.deleteGamePublisher = deleteGamePublisher;
+  public constructor({ sessionGetter, getGame, gameDAO }: Deps) {
+    this.sessionGetter = sessionGetter;
+    this.getGame = getGame;
     this.gameDAO = gameDAO;
   }
 
   public async execute(id: string): Promise<void> {
-    await this.gameDAO.delete(id);
+    const { uid } = await this.sessionGetter.getSession();
 
-    this.deleteGamePublisher.notifyDeleteGame(id);
+    const game = await this.getGame.execute(id);
+
+    if (!game)
+      throw new NotFoundError({
+        metadata: { entity: 'Game', prop: 'id', value: id },
+      });
+
+    if (game.uid !== uid)
+      throw new ForbiddenError({
+        metadata: { tried: 'delete a game of another user' },
+      });
+
+    await this.gameDAO.delete(id);
   }
 }
 
 type Deps = {
-  deleteGamePublisher: DeleteGameObserver.Publisher;
+  sessionGetter: SessionGetterProtocol;
+  getGame: IGetGameUsecase;
   gameDAO: IGameDAO;
 };
